requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==4.0.5
djangorestframework==3.13.1
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
